requirements_llm.txt
# LLM & Reasoning Engine Requirements
# Core dependencies for Member 1 responsibilities

# LLM and API Integration
groq>=0.4.1
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0

# Prompt Engineering and Templates
jinja2>=3.1.2
pydantic>=2.0.0

# RAG System
chromadb>=0.4.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Vector Embeddings
openai>=1.0.0  # For embeddings API
tiktoken>=0.5.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# JSON Schema and Validation
jsonschema>=4.17.0
pyyaml>=6.0

# Logging and Monitoring
loguru>=0.7.0
rich>=13.0.0

# HTTP and API
requests>=2.28.0
httpx>=0.24.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
